+++ EntityFX.NetBenchamarks.Core/Generic/MathBenchmark.cs
@@ -10,7 +10,7 @@ namespace EntityFX.NetBenchmark.Core.Generic
             double R = 0;
 
             double li = 0;
-            for (long i = 0; i < Iterrations; li = i, i++)
+            for (int i = 0; i < Iterrations; li = i, i++)
             {
                 R += DoMath(i, li);
             }
